+++ phpcs-server/src/settings.ts
@@ -26,4 +26,5 @@ export interface PhpcsSettings {
 	phpcbfEnable: boolean;
 	phpcbfExecutablePath: string | null;
 	phpcbfOnSave: boolean;
+	phpcbfTimeout: number;
 }